# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FXREBATE PROJECT",
        "description": "REACT Progressive Web App / Stand on the shoulders of giants",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/regulators/create": {
            "get": {
                "tags": [
                    "Regulator"
                ],
                "summary": "Get form data for creating regulator",
                "operationId": "6b6ca2713fb0aeb4fa56fd565439cb24",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Create form endpoint"
                                        },
                                        "data": {
                                            "properties": {
                                                "statuses": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "countries": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
...
}